requirements.txt
azure-core==1.24.1
azure-storage-blob==12.12.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
lxml==4.9.0
msrest==0.7.0
ntlm-auth==1.5.0
oauthlib==3.2.0
pycparser==2.21
requests==2.28.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SharePlum==0.5.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
